claude-const-maximizer/backend/requirements.txt
# Core CrewAI dependencies
crewai>=0.150.0

# LLM providers
anthropic>=0.18.0
openai>=1.50.0
google-generativeai>=0.3.0
langchain-google-genai>=0.1.0

# LangChain (required for agents)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# HTTP requests for services bootstrapper
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Utilities
python-slugify>=8.0.0

# Vector databases
pinecone-client>=3.0.0

# Search tools
tavily-python>=0.3.0

